# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Activo Fijo API — CLT S.A.
  version: "1.0.0"
  description: |
    API REST para el Sistema de Gestión de Activos Fijos de CLT S.A.
    Arquitectura Hexagonal · FastAPI · PostgreSQL 15
  contact:
    name: CLT — Equipo de Desarrollo
    email: desarrollo@clt.com.py
servers:
  - url: http://localhost:8000
    description: Desarrollo local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MantenimientoCreate:
      type: object
      required: [activo_id, tipo, descripcion, tecnico, fecha_inicio]
      properties:
        activo_id: { type: string, format: uuid }
        tipo:
          type: string
          enum: [preventivo, correctivo, predictivo]
        descripcion: { type: string }
        tecnico: { type: string }
        proveedor: { type: string }
        fecha_inicio: { type: string, format: date }
        fecha_estimada_fin: { type: string, format: date }
        costo: { type: number, minimum: 0, default: 0 }
    MantenimientoUpdate:
      type: object
      properties:
        estado:
          type: string
          enum: [programado, en_proceso, completado, cancelado]
        fecha_fin_real: { type: string, format: date }
        costo: { type: number, minimum: 0 }
        descripcion: { type: string }
    MantenimientoResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        activo_id: { type: string, format: uuid }
        activo_codigo: { type: string }
        activo_nombre: { type: string }
        tipo: { type: string }
        descripcion: { type: string }
        tecnico: { type: string }
        proveedor: { type: string }
        fecha_inicio: { type: string, format: date }
        fecha_estimada_fin: { type: string, format: date }
        fecha_fin_real: { type: string, format: date }
        costo: { type: number }
        estado: { type: string }
        created_at: { type: string, format: date-time }

    # ── AuditLog ──────────────────────────────────────────
...
paths:
...
  /v1/mantenimientos:
    get:
      tags: [Mantenimientos]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20 } }
        - { name: estado, in: query, schema: { type: string } }
        - { name: activo_id, in: query, schema: { type: string, format: uuid } }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/MantenimientoResponse' }
                  total: { type: integer }
    post:
      tags: [Mantenimientos]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MantenimientoCreate' }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MantenimientoResponse' }
  /v1/mantenimientos/{id}:
    patch:
      tags: [Mantenimientos]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MantenimientoUpdate' }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MantenimientoResponse' }

  # ── Audit Logs ────────────────────────────────────────────
...